fix(RT916): add .float() safety net for BFloat16->numpy conversion
- Two .cpu().numpy() call sites (validate, inference_single_period) now go through .float() first

- Defense-in-depth: even if OPTIM_AMP somehow re-enables, predictions come out as float32 and won't trip numpy's BFloat16 unsupported dtype

- Companion to the earlier OPTIM_AMP=0 fix in pipeline.py

+++ RT916_SpikeFusionNet/src/rt916_spikefusionnet/core.py
@@ -453,7 +453,7 @@ def _validate(model, val_loader, scaler_y, device):
 
     with torch.no_grad():
         for batch_x, batch_y in val_loader:
-            pred = model(batch_x.to(device)).cpu().numpy()
+            pred = model(batch_x.to(device)).float().cpu().numpy()
             all_preds.append(pred)
             all_trues.append(batch_y.numpy())
 
@@ -725,7 +725,7 @@ def inference_single_period(period_name, test_data, truth_df=None):
     all_preds, all_trues = [], []
     with torch.no_grad():
         for batch_x, batch_y in test_loader:
-            pred = model(batch_x.to(device)).cpu().numpy()
+            pred = model(batch_x.to(device)).float().cpu().numpy()
             all_preds.append(pred)
             all_trues.append(batch_y.numpy())
